+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Change the `/:storage/:name` route to `/:storage/upload` in direct_upload plugin (janko-m)
+
 * Fix logger not being inherited in the logging plugin (janko-m)
 
 * Add delete_promoted plugin for deleting promoted files after record has been updated (janko-m)